update MIMXRT1060 SDK to 2.8.2
* update MIMXRT1060 SDK to 2.8.2
* wait before sending first telemetry
* Explicitly default to MQTT

+++ NXP/MIMXRT1060-EVK/lib/MIMXRT1060-evk/src/components/codec/fsl_codec_common.c
@@ -7,14 +7,13 @@
  */
 
 #include "fsl_codec_common.h"
-#include "fsl_codec_adapter.h"
 /*******************************************************************************
  * Definitions
  ******************************************************************************/
 /*! @brief codec play and record capability */
-#define GET_PLAY_CHANNEL_CAPABILITY(capability) (capability & 0xFFU)
-#define GET_PLAY_SOURCE_CAPABILITY(capability) (capability >> 8U)
-#define GET_RECORD_SOURCE_CAPABILITY(capability) (capability & 0x3FU)
+#define GET_PLAY_CHANNEL_CAPABILITY(capability)   (capability & 0xFFU)
+#define GET_PLAY_SOURCE_CAPABILITY(capability)    (capability >> 8U)
+#define GET_RECORD_SOURCE_CAPABILITY(capability)  (capability & 0x3FU)
 #define GET_RECORD_CHANNEL_CAPABILITY(capability) (capability >> 6U)
 /*******************************************************************************
  * Variables
@@ -100,7 +99,7 @@ status_t CODEC_ModuleControl(codec_handle_t *handle, codec_module_ctrl_cmd_t cmd
             return kStatus_CODEC_NotSupport;
     }
 
-    return HAL_CODEC_ModuleControl(handle, cmd, data);
+    return HAL_CODEC_ModuleControl(handle, (uint32_t)cmd, data);
 }
 
 /*!
@@ -167,7 +166,7 @@ status_t CODEC_SetPower(codec_handle_t *handle, codec_module_t module, bool powe
         return kStatus_CODEC_NotSupport;
     }
 
-    return HAL_CODEC_SetPower(handle, module, powerOn);
+    return HAL_CODEC_SetPower(handle, (uint32_t)module, powerOn);
 }
 
 /*!
